refactor namespace cleaning
