+++ examples/foobar.c
@@ -1,4 +1,5 @@
 #include "../mpc.h"
+#include <stdlib.h>
 
 int main(int argc, char** argv) {
