+++ mpc.c
@@ -1,12 +1,4 @@
-#include "mpc.h"
-#ifndef va_copy
-# ifndef __va_copy
-#  define va_copy(dest, src) memcpy(&(dest), &(src), sizeof(va_list))
-# else
-#  define va_copy __va_copy
-# endif
-#endif
-
+#include "hidden/mpc_internal.h"
 
 /*
 ** State Type
@@ -64,49 +56,6 @@ static mpc_state_t mpc_state_new(void) {
 ** to parse for all input methods.
 **
 */
-
-enum {
-  MPC_INPUT_STRING = 0,
-  MPC_INPUT_FILE   = 1,
-  MPC_INPUT_PIPE   = 2
-};
-
-enum {
-  MPC_INPUT_MARKS_MIN = 32
-};
-
-enum {
-  MPC_INPUT_MEM_NUM = 512
-};
-
-typedef struct {
-  char mem[64];
-} mpc_mem_t;
-
-typedef struct {
-
-  int type;
-  char *filename;
-  mpc_state_t state;
-
-  char *string;
-  char *buffer;
-  FILE *file;
-
-  int suppress;
-  int backtrack;
-  int marks_slots;
-  int marks_num;
-  mpc_state_t *marks;
-
-  char *lasts;
-  char last;
-
-  size_t mem_index;
-  char mem_full[MPC_INPUT_MEM_NUM];
-  mpc_mem_t mem[MPC_INPUT_MEM_NUM];
-
-} mpc_input_t;
 
 static mpc_input_t *mpc_input_new_string(const char *filename, const char *string) {
 
@@ -890,92 +839,6 @@ static mpc_err_t *mpc_err_merge(mpc_input_t *i, mpc_err_t *x, mpc_err_t *y) {
 ** Parser Type
 */
 
-enum {
-  MPC_TYPE_UNDEFINED  = 0,
-  MPC_TYPE_PASS       = 1,
-  MPC_TYPE_FAIL       = 2,
-  MPC_TYPE_LIFT       = 3,
-  MPC_TYPE_LIFT_VAL   = 4,
-  MPC_TYPE_EXPECT     = 5,
-  MPC_TYPE_ANCHOR     = 6,
-  MPC_TYPE_STATE      = 7,
-
-  MPC_TYPE_ANY        = 8,
-  MPC_TYPE_SINGLE     = 9,
-  MPC_TYPE_ONEOF      = 10,
-  MPC_TYPE_NONEOF     = 11,
-  MPC_TYPE_RANGE      = 12,
-  MPC_TYPE_SATISFY    = 13,
-  MPC_TYPE_STRING     = 14,
-
-  MPC_TYPE_APPLY      = 15,
-  MPC_TYPE_APPLY_TO   = 16,
-  MPC_TYPE_PREDICT    = 17,
-  MPC_TYPE_NOT        = 18,
-  MPC_TYPE_MAYBE      = 19,
-  MPC_TYPE_MANY       = 20,
-  MPC_TYPE_MANY1      = 21,
-  MPC_TYPE_COUNT      = 22,
-
-  MPC_TYPE_OR         = 23,
-  MPC_TYPE_AND        = 24,
-
-  MPC_TYPE_CHECK      = 25,
-  MPC_TYPE_CHECK_WITH = 26,
-
-  MPC_TYPE_SOI        = 27,
-  MPC_TYPE_EOI        = 28,
-
-  MPC_TYPE_SEPBY1     = 29
-};
-
-typedef struct { char *m; } mpc_pdata_fail_t;
-typedef struct { mpc_ctor_t lf; void *x; } mpc_pdata_lift_t;
-typedef struct { mpc_parser_t *x; char *m; } mpc_pdata_expect_t;
-typedef struct { int(*f)(char,char); } mpc_pdata_anchor_t;
-typedef struct { char x; } mpc_pdata_single_t;
-typedef struct { char x; char y; } mpc_pdata_range_t;
-typedef struct { int(*f)(char); } mpc_pdata_satisfy_t;
-typedef struct { char *x; } mpc_pdata_string_t;
-typedef struct { mpc_parser_t *x; mpc_apply_t f; } mpc_pdata_apply_t;
-typedef struct { mpc_parser_t *x; mpc_apply_to_t f; void *d; } mpc_pdata_apply_to_t;
-typedef struct { mpc_parser_t *x; mpc_dtor_t dx; mpc_check_t f; char *e; } mpc_pdata_check_t;
-typedef struct { mpc_parser_t *x; mpc_dtor_t dx; mpc_check_with_t f; void *d; char *e; } mpc_pdata_check_with_t;
-typedef struct { mpc_parser_t *x; } mpc_pdata_predict_t;
-typedef struct { mpc_parser_t *x; mpc_dtor_t dx; mpc_ctor_t lf; } mpc_pdata_not_t;
-typedef struct { int n; mpc_fold_t f; mpc_parser_t *x; mpc_dtor_t dx; } mpc_pdata_repeat_t;
-typedef struct { int n; mpc_parser_t **xs; } mpc_pdata_or_t;
-typedef struct { int n; mpc_fold_t f; mpc_parser_t **xs; mpc_dtor_t *dxs;  } mpc_pdata_and_t;
-typedef struct { int n; mpc_fold_t f; mpc_parser_t *x; mpc_parser_t *sep; } mpc_pdata_sepby1;
-
-typedef union {
-  mpc_pdata_fail_t fail;
-  mpc_pdata_lift_t lift;
-  mpc_pdata_expect_t expect;
-  mpc_pdata_anchor_t anchor;
-  mpc_pdata_single_t single;
-  mpc_pdata_range_t range;
-  mpc_pdata_satisfy_t satisfy;
-  mpc_pdata_string_t string;
-  mpc_pdata_apply_t apply;
-  mpc_pdata_apply_to_t apply_to;
-  mpc_pdata_check_t check;
-  mpc_pdata_check_with_t check_with;
-  mpc_pdata_predict_t predict;
-  mpc_pdata_not_t not;
-  mpc_pdata_repeat_t repeat;
-  mpc_pdata_and_t and;
-  mpc_pdata_or_t or;
-  mpc_pdata_sepby1 sepby1;
-} mpc_pdata_t;
-
-struct mpc_parser_t {
-  char *name;
-  mpc_pdata_t data;
-  char type;
-  char retained;
-};
-
 static mpc_val_t *mpcf_input_nth_free(mpc_input_t *i, int n, mpc_val_t **xs, int x) {
   int j;
   for (j = 0; j < n; j++) { if (j != x) { mpc_free(i, xs[j]); } }
@@ -1045,18 +908,6 @@ static void mpc_parse_dtor(mpc_input_t *i, mpc_dtor_t d, mpc_val_t *x) {
   if (d == free) { mpc_free(i, x); return; }
   d(mpc_export(i, x));
 }
-
-enum {
-  MPC_PARSE_STACK_MIN = 4
-};
-
-#define MPC_SUCCESS(x) r->output = x; return 1
-#define MPC_FAILURE(x) r->error = x; return 0
-#define MPC_PRIMITIVE(x) \
-  if (x) { MPC_SUCCESS(r->output); } \
-  else { MPC_FAILURE(NULL); }
-
-#define MPC_MAX_RECURSION_DEPTH 1000
 
 static mpc_result_t *mpc_grow_results(mpc_input_t *i, int j, mpc_result_t *results_stk, mpc_result_t *results){
   mpc_result_t *tmp_results = results;
@@ -3433,13 +3284,6 @@ mpc_parser_t *mpca_total(mpc_parser_t *a) { return mpc_total(a, (mpc_dtor_t)mpc_
 **             | "(" <grammar> ")"
 */
 
-typedef struct {
-  int create_new;
-  int parsers_num;
-  mpc_parser_t **parsers;
-  int flags;
-} mpca_grammar_st_t;
-
 static mpc_val_t *mpcaf_grammar_or(int n, mpc_val_t **xs) {
   (void) n;
   if (xs[1] == NULL) { return xs[0]; }
@@ -3682,12 +3526,6 @@ mpc_parser_t *mpca_grammar(int flags, const char *grammar, ...) {
   free(st.parsers);
   return res;
 }
-
-typedef struct {
-  char *ident;
-  char *name;
-  mpc_parser_t *grammar;
-} mpca_stmt_t;
 
 static mpc_val_t *mpca_stmt_afold(int n, mpc_val_t **xs) {
   mpca_stmt_t *stmt = malloc(sizeof(mpca_stmt_t));
