+++ mpc.h
@@ -14,15 +14,7 @@
 extern "C" {
 #endif
 
-#include <stdlib.h>
 #include <stdio.h>
-#include <stdarg.h>
-#include <string.h>
-#include <math.h>
-#include <errno.h>
-#include <ctype.h>
-#include <fcntl.h>
-#include <unistd.h>
 
 /*
 ** State Type
